Main

Types

data Command

Type for Player Actions

Constructors

Open Pos
Open command with position.

Flag Pos
Flag command with position.

Invalid String
Invalid command for unknown player input.

Game functions

parseInput

Arguments

:: MonadReader GameEnv m

=> String
Player input.

-> m Command
Game command.

Parse player input into a Player Action Command.

gameLoop :: (MonadReader GameEnv m, MonadIO m, MonadState GameState m) => m ()

Game loop that contiually requets Player commands and draws the updated game board after each valid command until the game completes.

parseArgs :: [String] -> Maybe GameEnv

Create a GameEnv from opitonal arguments.

promptUser

Arguments

:: (MonadState GameState m, MonadIO m)

=> m String
Player input.

Draw the current board state and prompt Player for command.

startGame :: (MonadReader GameEnv m, MonadIO m, MonadState GameState m) => m ()

Draw intial board user Player settings and get Player command. Start the game loop that repeats prompting for Player commands and rendering game board.

defaultDifficulty :: Difficulty

Default difficulty settings. Mined cells between 20% and 10% of total cells.

defaultEnvironment :: GameEnv

Default game board 20x20 with default difficulty.

main :: IO ()

Create a GameEnv from optional arguments or with defaults. Start the game loop and run until game completes.