MineSweeper

Description

Game Types and functions for a MineSweeper Game.

Game actions

newGameBoard :: (MonadReader GameEnv m, MonadState GameState m) => m ()

Create a new unmined board using Players settings.

mineBoard

Arguments

:: (MonadReader GameEnv m, MonadIO m, MonadState GameState m)

=> Pos
Position where user first opened a cell.

-> m ()

Take a new board and randomly mine it based on the players settings for difficulty. The player would have opened one cell and we make sure we don't mine directly adjacent to that cell.

Game state

updateBoardState

Arguments

:: (MonadState GameState m, MonadReader GameEnv m)

=> Board
Board to be udpated in GameState.

-> m ()

Update the game board in state with evaluated game status and statistics.

User actions

openCell

Arguments

:: (MonadReader GameEnv m, MonadState GameState m)

=> Pos
Cell position requested to opened.

-> m ()

flagCell

Arguments

:: (MonadReader GameEnv m, MonadState GameState m)

=> Pos
Cell position requestd to be flagged.

-> m ()

User action to flag a cell.